added specific integration for mct

+++ examples/examples_1D/01_linearadvection/mcT_adv_data.py
@@ -100,7 +100,7 @@ class Data():
         for ii in range(setup.num_train):
             data_train[ii,...] = self._load(self.next_sim())
 
-        data_test = np.zeros((setup.num_test,5,setup.nt*setup.test_ratio+1,setup.nx_fine,setup.ny_fine,setup.nz_fine))
+        data_test = np.zeros((setup.num_test,5,int(setup.nt*setup.test_ratio)+1,setup.nx_fine,setup.ny_fine,setup.nz_fine))
         for ii in range(setup.num_test):
             data_test[ii,...] = self._load(self.next_sim())
